requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-environ==0.9.0
django-js-asset==2.0.0
django-modeltranslation==0.17.5
gunicorn==20.1.0
Pillow==9.0.1
pip==22.0.3
python-decouple==3.6
setuptools==60.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
wheel==0.37.1
whitenoise==6.0.0
